refactor coordinate display to react, add top level react component
