+++ package.json
@@ -84,6 +84,7 @@
     "cheerio": "^1.0.0-rc.2",
     "chromedriver": "2.25.1",
     "clean-webpack-plugin": "^0.1.19",
+    "copy-webpack-plugin": "^4.5.2",
     "cross-env": "^5.2.0",
     "css-hot-loader": "^1.4.1",
     "css-loader": "^1.0.0",
@@ -199,5 +200,8 @@
     "transformIgnorePatterns": [
       "node_modules/(?!(ol)/)"
     ]
+  },
+  "prettier": {
+    "singleQuote": true
   }
 }
