+++ web/index.html
@@ -46,64 +46,6 @@ Licensed under the NASA Open Source Agreement, Version 1.3 (http://opensource.gs
 
 <body>
   <div data-role="page" id="app">
-    <div data-role="content">
-      <ul id="wv-toolbar">
-        <li id="wv-link-button" class="wv-toolbar-button"></li>
-        <li id="wv-proj-button" class="wv-toolbar-button"></li>
-        <li id="wv-image-button" class="wv-toolbar-button"></li>
-        <li id="wv-info-button" class="wv-toolbar-button wv-status-hide"></li>
-      </ul>
-      <section id="wv-sidebar">
-      </section>
-      <div id="wv-map" class="wv-map"></div>
-      <div id="eventsHolder"></div>
-      <div id="imagedownload"></div>
-      <div id="dlMap"></div>
-
-      <div id="timewheels" style="display:none"></div>
-
-      <section id="timeline" class="timeline-inner clearfix" style="display: none">
-        <div id="timeline-header">
-          <div id="date-selector-main"></div>
-          <div id="zoom-buttons-group">
-            <div id="zoom-btn-container">
-              <span id="current-zoom" class="zoom-btn zoom-btn-active" data-zoom="3">Days</span>
-              <div class="wv-tooltip">
-                <div id="timeline-zoom" class="timeline-zoom">
-                  <span id="zoom-years" class="zoom-btn zoom-btn-inactive zoom-years" data-zoom="1">Years</span>
-                  <span id="zoom-months" class="zoom-btn zoom-btn-inactive zoom-months" data-zoom="2">Months</span>
-                  <span id="zoom-days" class="zoom-btn zoom-btn-inactive zoom-days" data-zoom="3">Days</span>
-                  <span id="zoom-minutes" class="zoom-btn zoom-btn-inactive zoom-minutes" data-zoom="4">Minutes</span>
-                </div>
-              </div>
-            </div>
-            <div class="button-action-group" id="left-arrow-group" title="Click and hold to animate backwards">
-              <svg id="timeline-svg" width="24" height="30">
-                <path d="M 10.240764,0 24,15 10.240764,30 0,30 13.759236,15 0,0 10.240764,0 z" class="arrow"></path>
-              </svg>
-            </div>
-            <div class="button-action-group" id="right-arrow-group" title="Click and hold to animate forwards">
-              <svg width="24" height="30">
-                <path d="M 10.240764,0 24,15 10.240764,30 0,30 13.759236,15 0,0 10.240764,0 z" class="arrow"></path>
-              </svg>
-            </div>
-          </div>
-          <div class="button-action-group animate-button" id="animate-button" title="Set up animation">
-            <i id="wv-animate" class="fa fa-video-camera wv-animate"></i>
-          </div>
-
-        </div>
-        <div id="timeline-footer">
-          <div id="wv-animation-widet-case"> </div>
-        </div>
-        <div id="timeline-hide">
-          <svg class="hamburger" width="10" height="9">
-            <path d="M 0,0 0,1 10,1 10,0 0,0 z M 0,4 0,5 10,5 10,4 0,4 z M 0,8 0,9 10,9 10,8 0,8 z"></path>
-          </svg>
-        </div>
-      </section>
-
-    </div>
   </div>
   <section id="layer-modal" class="layer-modal">
     <header id="layer-modal-header" class="layer-modal-header"></header>
