+++ web/js/util/util.test.js
@@ -320,19 +320,19 @@ describe('formatDMS', () => {
   let tests = [{
     name: 'zero',
     dd: [0, 0],
-    str: [' 0&deg;00\'00"N', '  0&deg;00\'00"E']
+    str: [' 0°00\'00"N', '  0°00\'00"E']
   }, {
     name: 'positive',
     dd: [12.582222, 12.582222],
-    str: ['12&deg;34\'55"N', ' 12&deg;34\'55"E']
+    str: ['12°34\'55"N', ' 12°34\'55"E']
   }, {
     name: 'negative',
     dd: [-12.582222, -12.582222],
-    str: ['12&deg;34\'55"S', ' 12&deg;34\'55"W']
+    str: ['12°34\'55"S', ' 12°34\'55"W']
   }, {
     name: 'max',
     dd: [89.999997, 179.999997],
-    str: ['89&deg;59\'59"N', '179&deg;59\'59"E']
+    str: ['89°59\'59"N', '179°59\'59"E']
   }];
 
   tests.forEach((t) => {
@@ -349,19 +349,19 @@ describe('formatDM', () => {
   let tests = [{
     name: 'zero',
     dd: [0, 0],
-    str: [' 0&deg;00.000\'N', '  0&deg;00.000\'E']
+    str: [' 0°00.000\'N', '  0°00.000\'E']
   }, {
     name: 'positive',
     dd: [12.582222, 12.582222],
-    str: ['12&deg;34.933\'N', ' 12&deg;34.933\'E']
+    str: ['12°34.933\'N', ' 12°34.933\'E']
   }, {
     name: 'negative',
     dd: [-12.582222, -12.582222],
-    str: ['12&deg;34.933\'S', ' 12&deg;34.933\'W']
+    str: ['12°34.933\'S', ' 12°34.933\'W']
   }, {
     name: 'max',
     dd: [89.999997, 179.999997],
-    str: ['89&deg;59.999\'N', '179&deg;59.999\'E']
+    str: ['89°59.999\'N', '179°59.999\'E']
   }];
 
   tests.forEach((t) => {
